requirements.txt
streamlit==1.35.0
playwright==1.44.0
beautifulsoup4==4.12.3
openai==1.30.1
google-genai==0.5.0
pandas==2.2.2
openpyxl==3.1.2
ollama==0.2.1
